authors titles mark up
I'm uploading an edited version of riddle-arnold_for_web.csv renamed as riddle-arnold_for_web_authors_titles.csv.  The basic structure has not changed, but I've added <author> and <title> tags, as a step possibly in the direction of xml. Also, I've pruned the 'latin' markup to include less of the English text. Obviously, there is plenty left to do..  Caution: I've done some work on synchronizing authors and titles with Lewis & Short but again, this work is definitely not complete.